add common script for utils functions
